@@ -1,23 +1,21 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>demo-https</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-    <name>demo-https</name>
-  <description>Demo project for Spring Boot</description>
-
   <parent>
     <groupId>com.xkcoding</groupId>
-    <artifactId>spring-boot-demo</artifactId>
+    <artifactId>demo-others</artifactId>
     <version>1.0.0-SNAPSHOT</version>
   </parent>
 
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>demo-others-https</artifactId>
+  <version>0.0.1-SNAPSHOT</version>
+  <name>demo-others-https</name>
+  <description>Demo project for Spring Boot</description>
+
   <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <java.version>1.8</java.version>
+    <java.version>17</java.version>
   </properties>
 
   <dependencies>
@@ -25,6 +23,7 @@
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-web</artifactId>
     </dependency>
+
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-test</artifactId>
@@ -33,6 +32,7 @@
   </dependencies>
 
   <build>
+    <finalName>demo-others-https</finalName>
     <plugins>
       <plugin>
         <groupId>org.springframework.boot</groupId>
