@@ -1,10 +1,10 @@
 package com.xkcoding.https;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SpringBootDemoHttpsApplicationTests {
+class HttpsApplicationTests {
 
     @Test
     void contextLoads() {
