@@ -25,9 +25,9 @@ public class HttpsConfig {
     public Connector connector() {
         Connector connector = new Connector("org.apache.coyote.http11.Http11NioProtocol");
         connector.setScheme("http");
-        connector.setPort(80);
+        connector.setPort(8080);
         connector.setSecure(false);
-        connector.setRedirectPort(443);
+        connector.setRedirectPort(8443);
         return connector;
     }
 
